requirements.txt
tensorflow==2.12.0
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
gdown
torch
diffusers
rembg 
pywhatkit 
pyautogui 
pillow
auth-token

-e .